ajout mofidication de nom de medicament

+++ remove_and_modify.php
@@ -2,14 +2,41 @@
 
 include("conn.php");
 
-$id = $_GET['id'];
+if ($_POST['modify']){
+    $old_name = strval($_POST['to_modify']);
+    $new_name = $_POST['new_name_medic'];
 
-$sql = "DELETE FROM `tableau` WHERE `tid` = $id";
+    $sql = "SELECT * FROM `tableau` WHERE `medicament` LIKE '%".$old_name."%'";
+    $result = mysqli_query($conn, $sql);
 
-if (!mysqli_query($conn, $sql)) {
-    die("Error: " . mysqli_error($conn));
+    while ($rows = mysqli_fetch_array($result)){
+        $id = $rows['tid'];
+        $sql =" UPDATE `tableau` SET `medicament` = '".$new_name."' WHERE `tableau`.`tid` = $id";
+
+        if(!mysqli_query($conn, $sql)){
+            die("Error: ".mysqli_error($conn));
+        }
+    }
+
+    echo $old_name;
+    mysqli_close($conn);
+
+    die("<script>window.location.href= 'medicatime.php'</script>");
+
+}else{
+    $id = $_GET['id'];
+
+    $sql = "DELETE FROM `tableau` WHERE `tid` = $id";
+
+    if (!mysqli_query($conn, $sql)) {
+        die("Error: " . mysqli_error($conn));
+    }
+
+    mysqli_close($conn);
+
+    die("<script>window.location.href= 'medicatime.php'</script>");
 }
 
-mysqli_close($conn);
 
-die("<script>window.location.href= 'medicatime.php'</script>");
+
+
